refactor(backend): extract DB operation to DumpsRepository
