+++ backend/app/generic/apis/default_api.py
@@ -16,10 +16,9 @@ from fastapi import (
     Security,
     status,
 )
-from sqlalchemy.orm import Session
 
-from app.db import select_dumps
-from app.deps import get_db
+from app.db.repositories.dumps import DumpsRepository
+from app.dependencies.database import get_repo
 from app.generic.models.bad_request import BadRequest
 from app.generic.models.dump import Dump
 from app.generic.models.dump_elements import DumpElements
@@ -48,9 +47,8 @@ async def dumps_get(
     cursor: str = Query(None, description=""),
     rows: int = Query(None, description="", ge=0, le=100),
     token_main_security_scheme: TokenModel = Security(get_token_main_security_scheme),
-    db: Session = Depends(get_db),
+    dumps_repo: DumpsRepository = Depends(get_repo(DumpsRepository)),
 ) -> DumpResults:
     """Returns available dumps"""
     # pylint: disable=unused-argument
-    dumps = select_dumps(db=db)
-    return DumpResults(dumps=dumps)
+    return DumpResults(dumps=dumps_repo.list_all_dumps())
