+++ backend/app/manager.py
@@ -4,8 +4,8 @@ from datetime import datetime, timezone
 
 import typer
 
-from .db import SessionLocal
 from .models import Dump, DumpElement
+from .tasks import create_session_local
 
 db_group = typer.Typer()
 
@@ -15,7 +15,7 @@ def seed_basic():
     """
     Seed with a basic data-set
     """
-    with SessionLocal() as session:
+    with create_session_local()() as session:
         dump = Dump(
             name="openaire_1",
             created_at=datetime.now(timezone.utc),
@@ -39,7 +39,7 @@ def seed_oag_1():
     Seed with oag-1 data-set
     """
     s3_prefix = "https://ess-mock-dumps.s3.cloud.cyfronet.pl"
-    with SessionLocal() as session:
+    with create_session_local()() as session:
         dump = Dump(
             name="oag_1",
             created_at=datetime.now(timezone.utc),
@@ -70,7 +70,7 @@ def seed_oag_2():
     Seed with oag-2 data-set
     """
     s3_prefix = "https://ess-mock-dumps.s3.cloud.cyfronet.pl"
-    with SessionLocal() as session:
+    with create_session_local()() as session:
         dump = Dump(
             name="oag_2",
             created_at=datetime.now(timezone.utc),
